Create new string when appending /
The reason is that sometimes, you will get a default value here.
If this is the case, this value is shared with all resources of
that type. If you modify it for one, you modify it for all. This
will break in the next version of Chef because the default value
will be frozen.

See https://github.com/chef/chef/commit/705e22b975196baa6decf23f9c3a7e676d6aefc4