Add support for CentOS 8
CentOS 8 switched to using dnf instead of yum so the yum-cache-reload ruby_block
needs to be updated to use
``Chef::Provider::Package::Dnf::PythonHelper.instance.restart`` on CentOS 8.

In addition, CentOS 8 removes the pygpgme as dnf does not require it so let's
not emit warnings about it on that platform.

Signed-off-by: Lance Albertson <lance@osuosl.org>